requirements.txt
streamlit>=1.20.0
python-dotenv>=1.0.0
azure-search-documents>=11.5.0
azure-ai-openai>=1.0.0
azure-core>=1.30.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
langchain-openai>=0.0.9
requests>=2.28.0
